eventually split federation
