+++ packages/nestjs/package.json
@@ -32,20 +32,6 @@
         "default": "./dist/esm/index.js"
       }
     },
-    "./*": {
-      "require": {
-        "types": "./dist/typings/*.d.cts",
-        "default": "./dist/cjs/*.js"
-      },
-      "import": {
-        "types": "./dist/typings/*.d.ts",
-        "default": "./dist/esm/*.js"
-      },
-      "default": {
-        "types": "./dist/typings/*.d.ts",
-        "default": "./dist/esm/*.js"
-      }
-    },
     "./package.json": "./package.json"
   },
   "typings": "dist/typings/index.d.ts",
